don't reuse the phrase Argonaut Scheduling
Argonaut scheduling is a distinct IG: http://www.fhir.org/guides/argonaut/scheduling/ that also uses Subscriptions, but with an extension. The overall approach laid out in this file is highly applicable to open scheduling pubsub, but is different from what's defined in the Argo scheduling ig. Because the two use-cases and proposed (and as of yet non-standardized) use of Subscriptions are so intertwined, we should be careful to not confuse them.